Fread cleanup (#653)
* Make sure that all ctx_parser* functions set NA values on failure

* Avoid reallocating FieldParseContext on every iteration

* Synchronize sections 11 and 12 of freadMain

* Synchronize Field() parser

* Synchronize ctx_Field and countfields functions

* Update (almost) all parsers

* Update boolean field parsers

* Clean up temporary variable/function names

* Final cleanup in fread

* Fix warnings

* Rename fread.c->.cc

* Fix a bug in omnibus test

* Add test for issue 643

+++ c/csv/reader.h
@@ -107,7 +107,7 @@ class GenericReader
     MemoryBuffer* mbuf;
     size_t offset;
     size_t offend;
-    size_t line;
+    int64_t line;
     int32_t fileno;
     int : 32;
